fix(lint): clear golangci-lint findings unmasked on main (part 1)
golangci-lint now runs to completion on main (libudev fix), exposing a
pre-existing backlog that the PR path hides via --new-from-rev. Remove
genuinely dead code and fix real findings; no gate weakening, no test
skips, no config changes.

+++ pkg/provider_daemon/backendprofile/contracts.go
@@ -261,7 +261,7 @@ func Validate(profile CloudProfile, graph DesiredResourceGraph) error {
 		used.StorageGiB += resource.Budget.StorageGiB
 		used.CostMinorUnits += resource.Budget.CostMinorUnits
 	}
-	if uint32(len(resources)) > profile.Quotas.Resources || used.VCPUs > profile.Quotas.VCPUs || used.MemoryMiB > profile.Quotas.MemoryMiB || used.StorageGiB > profile.Quotas.StorageGiB {
+	if uint64(len(resources)) > uint64(profile.Quotas.Resources) || used.VCPUs > profile.Quotas.VCPUs || used.MemoryMiB > profile.Quotas.MemoryMiB || used.StorageGiB > profile.Quotas.StorageGiB {
 		return errors.New("resource graph exceeds profile quota")
 	}
 	if used.CostMinorUnits > profile.CostCeiling.MinorUnits {
